+++ app/models/unidom/geo/concerns/as_region.rb
@@ -1,0 +1,15 @@
+module Unidom::Geo::Concerns::AsRegion
+
+  extend ActiveSupport::Concern
+
+  included do |includer|
+
+    has_many :locations, class_name: 'Unidom::Geo::Location'
+    has_many :locatings, through:    :locations,              source: :locatings
+
+  end
+
+  module ClassMethods
+  end
+
+end