+++ CHANGELOG.md
@@ -20,3 +20,7 @@
 
 ## v1.0.1
 1. Improved the Ruby Gem Specification to depend on [unidom-common](https://github.com/topbitdu/unidom-common) v1.0
+
+## v1.1
+1. As Located concern
+2. As Region concern